添加时间判断

+++ src/main/kotlin/Data.kt
@@ -6,7 +6,7 @@ import net.mamoe.mirai.console.data.value
 object Data : AutoSavePluginData("Data") {
     var access_token: String by value()
 
-    var access_date: Int by value()
+    var access_date: Long by value()
 
     val cookie: String by value()
 }
